Added TUID User model.
Objects of this model can't be changed or created in the admin interface.
When TUID_CREATE_USER is set to True, objects of this model will be created on first login of a TUID User and are updated to the new values returned by CAS everytime the user logs in again.